Lots of fixes on calculate possible moves...
Which tells me that we might need to think a little bit harder about how
to manage that.  The existing functionality does a great job of
determining what squares are available to a piece based upon it's basic
movement type, but it fails exclude all squares appropriately when given
reason to, (i.e., if some square in between the piece and it's
destination is occupied, either by its own piece or an opponent's).

I think the solution will be to move away from the looping constructs