New Piece class
And a whole new mechanism for determining available/legal moves.
Available moves are determined on the Piece class, limited only by the
type of piece making the move, and their position on the board.  This
list is further filtered by the Board, with respect to the layout of
other pieces present.

This is a much cleaner interface for determining move legality, and it
also gave me the opportunity to cleanly write a method to help determine
'legality based upon the state of the board'.  That method is a little
larger than I want, but we can refactor it down to various pieces later
-- the code is relatively clean at least.

There are still some bugs in the move parser, but as it stands, we've
gotten further through the sample game than I ever have, so it's
definitely progress in the right direction