Basic parsing for the win
This is an important commit -- I was able to simplify the parsing rules
for PGN files, making it a lot more straightforward to figure out moves.
With that said, I want to be able to support the display of comments, so
my next trick will be to find a sample game that includes them.  Not
sure that I will ever support variations, but that's something to think
about too.