Better move recording from PGN parsing.
Still not quite there with registering comments in PGN. _Extracting_ the
comments as entities unto themselves is easy enough -- the problem comes
in trying to assign them to the appropriate move.  I have to consider
what the appropriate regex would look like...all of my attempts so far
have grabbed too much _beyond_ the content of the comment.