Much more intelligent way of parsing movetext
This one relying upon the fact that all moves _are_ numbered...we split
on the numbers, and can perform additional parsing from there.

I also tried to do a little bit of annotation parsing, but I'm falling a
bit short here.  For one thing, I'm assuming one comment per _move_ when
in fact, it's really one comment per _ply_.  There also seems to be
support for the idea of a 'pregame' commentary which precedes the
movetext block entirely.

So, tasks in front of us for the PGN parser, (which is really a distinct
project from the chessboard itself), are to properly parse commentary
text.

Another quick note on that subject, just so I'm aware of all caveats: It
would appear that the movetext parser _might_ be picking up on the
movetext listed in variations...which would be a bad thing, and needs to
be looked in to.  I think the solution there is going to have to be to
extract comments _before_ parsing movetext, thereby ensuring that the
comments aren't considered.